requirements.txt
Django==1.11
django-birthday==0.1.2
djangorestframework==3.7.7
Faker==0.8.11
Pillow==5.0.0
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
text-unidecode==1.1
click==6.7
django-birthday==0.1.2
djangorestframework==3.7.7
Faker==0.8.11
Pillow==5.0.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2018.3
six==1.11.0
pytz==2018.3
